+++ tests/utilities/indexer/test_index_builder.cpp
@@ -39,8 +39,8 @@ void run_coro(Fn&& fn) {
 TEST_SUITE("IndexBuilder") {
     TEST_CASE("Build checkpoint-only index") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         auto config = IndexBuildConfig::for_file(gz_file).with_manifest(false);
 
@@ -100,8 +100,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Build with bloom") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         auto config = IndexBuildConfig::for_file(gz_file)
 
@@ -128,8 +128,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Build with manifest") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         auto config = IndexBuildConfig::for_file(gz_file)
 
@@ -156,8 +156,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Build with bloom and manifest") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         auto config = IndexBuildConfig::for_file(gz_file)
 
@@ -185,8 +185,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Skip if already indexed") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         auto config = IndexBuildConfig::for_file(gz_file)
 
@@ -218,8 +218,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Force rebuild") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         auto config_normal = IndexBuildConfig::for_file(gz_file)
 
@@ -258,7 +258,8 @@ TEST_SUITE("IndexBuilder") {
     TEST_CASE("Result has correct line count") {
         const std::size_t n_idx_lines = valgrind_scale(1000, 10);
         TestEnvironment env(n_idx_lines);
-        std::string gz_file = env.create_dft_test_gzip_file(n_idx_lines);
+        std::string gz_file =
+            env.create_dft_test_gzip_file(static_cast<int>(n_idx_lines));
 
         auto config = IndexBuildConfig::for_file(gz_file).with_manifest(false);
 
@@ -278,8 +279,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Incremental manifest add to existing index with bloom") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         // First build: checkpoint + bloom
         auto config1 = IndexBuildConfig::for_file(gz_file)
@@ -330,8 +331,8 @@ TEST_SUITE("IndexBuilder") {
 
     TEST_CASE("Skip when all requested features already exist") {
         TestEnvironment env(valgrind_scale(1000, 10));
-        std::string gz_file =
-            env.create_dft_test_gzip_file(valgrind_scale(1000, 10));
+        std::string gz_file = env.create_dft_test_gzip_file(
+            static_cast<int>(valgrind_scale(1000, 10)));
 
         // Build with bloom + manifest
         auto config1 = IndexBuildConfig::for_file(gz_file)
